Update desintegration.jl
Modelling and simulation of incompressible flow in a two-dimensional domain with buoyancy effects:
A comprehensive study using Navier-Stokes Equations. The mass transport equation contains a reaction rate term and initially contains only two concentrations that simulate two biochemical species from the ADM1 model. In this case, it simulates the disintegration stage in a simplified way. It is already a preview to continue the simulation with 24 biochemical species in its 19 biochemical processes of the ADM1 model.